ocl/requirements.txt
dash==2.16.1  
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
numpy==1.26.4
ollama==0.1.9
black==24.4.2  
pylint==3.1.0
python-dotenv==1.0.1
PyYAML==6.0.1